requirements.txt
Pillow
mysql-connector-python
opencv-python
dlib
numpy
